+++ nodalia-cards-2.1.3-alpha.3.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.1.3-alpha.5.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.5","contentSha256_12":"9f6202a967a2","file":"nodalia-cards-2.1.3-alpha.3.js","targetFile":"nodalia-cards-2.1.3-alpha.5.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.1.3-alpha.6.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.6","contentSha256_12":"9aa6153346bd","file":"nodalia-cards-2.1.3-alpha.3.js","targetFile":"nodalia-cards-2.1.3-alpha.6.js","fallbackFile":"nodalia-cards.js"};}