+++ nodalia-cards.manifest.js
@@ -1,6 +1,6 @@
 export default {
-  "pkgVersion": "2.1.3-alpha.5",
-  "contentSha256_12": "9f6202a967a2",
+  "pkgVersion": "2.1.3-alpha.6",
+  "contentSha256_12": "9aa6153346bd",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
@@ -8,16 +8,16 @@ export default {
     "nodalia-cards-2.1.3-alpha.2.js",
     "nodalia-cards-2.1.3-alpha.3.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.5.js",
+  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.6.js",
   "splitCoreSha256_12": "f2cb3da1dbe5",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.5.js",
-  "splitSuiteSha256_12": "45747c53c40e",
-  "editorFile": "nodalia-cards-editor-2.1.3-alpha.5.js",
-  "editorSha256_12": "6b4f68820f49"
+  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.6.js",
+  "splitSuiteSha256_12": "9ddfc2c7417f",
+  "editorFile": "nodalia-cards-editor-2.1.3-alpha.6.js",
+  "editorSha256_12": "910d4e2d8ad0"
 };
-export const pkgVersion = "2.1.3-alpha.5";
-export const contentSha256_12 = "9f6202a967a2";
+export const pkgVersion = "2.1.3-alpha.6";
+export const contentSha256_12 = "9aa6153346bd";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.5.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.5.js";
-export const editorFile = "nodalia-cards-editor-2.1.3-alpha.5.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.6.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.6.js";
+export const editorFile = "nodalia-cards-editor-2.1.3-alpha.6.js";